Refactor: Separate out prompt generation from model.
The logic was getting a bit too entangled and difficult to extend.

This breaks out the prompt generation into its a separate prompt_builder
module, as well as introduces Prompt classes (with both TextPrompt and
OpenAIPrompt as two implementations).

This also changes the model interfaces to take in a Prompt object
instead of `prompt_path`. `prompt_path` was only really needed for
logging and AIBinaryModel.

Ref: #62.
